requirements.txt
-r submodules/parafoil/requirements.txt
-r submodules/flyplot/requirements.txt

# aiogram 3x
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
pydantic==1.10.7
typing_extensions==4.5.0
yarl==1.9.2

# dotenv
python-dotenv==1.0.0